pyproject.toml
[project]
name = "legal-contract-parser"
version = "0.1.0"
description = "AI-powered legal contract parser with LLM extraction"
authors = [
    {name = "angad bashani",email = "angadbashani@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.121.3,<0.122.0)",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "pydantic (>=2.12.4,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pypdf2 (>=3.0.1,<4.0.0)",
    "pdfplumber (>=0.11.8,<0.12.0)",
    "openai (>=2.8.1,<3.0.0)",
    "anthropic (>=0.74.1,<0.75.0)",
    "langchain (>=1.0.8,<2.0.0)",
    "langchain-openai (>=1.0.3,<2.0.0)",
    "langgraph (>=1.0.3,<2.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "pymupdf (>=1.26.6,<2.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "chromadb (>=1.3.5,<2.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "langchain-text-splitters (>=1.0.0,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]

[tool.poetry]
packages = [{include = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.1"
pytest-asyncio = "^1.3.0"
httpx = "^0.28.1"
black = "^25.11.0"
flake8 = "^7.3.0"
mypy = "^1.18.2"
prometheus-client = "^0.23.1"

[tool.black]
line-length = 100
target-version = ['py312']  # Match your Python 3.12
include = '\.pyi?$'

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
asyncio_mode = "auto"
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[tool.mypy]
python_version = "3.12"  # Match your Python 3.12
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

